changed checkstyle suppression with more specific one. Removed code duplication with cluster authetnicator. Added new ctors for ClientInvocation with/out EventHandler. Moved HazelcastClientNotActive from impl to com.hazelcast.client package
