+++ hazelcast-client/src/main/java/com/hazelcast/client/impl/ClusterAuthenticator.java
@@ -1,0 +1,9 @@
+package com.hazelcast.client.impl;
+
+/**
+ * User: sancar
+ * Date: 09/01/15
+ * Time: 17:08
+ */
+public class ClusterAuthenticator {
+}
